+++ plugins/web-interface/deps.ts
@@ -1,3 +1,3 @@
 export { Application, Router } from "https://deno.land/x/oak@14.2.0/mod.ts"
-export { dirname, fromFileUrl } from "https://deno.land/std@0.221.0/path/mod.ts"
+export { dirname, fromFileUrl } from "@std/path"
 export { Bundlee } from "https://deno.land/x/bundlee@0.9.4/mod.ts"